+++ src/main/java/com/asg/shipchandling/deliverynote/entity/SalesDeliveryNoteHdr.java
@@ -8,7 +8,7 @@ import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.Generated;
 import org.hibernate.annotations.GenerationTime;
 import org.hibernate.annotations.UpdateTimestamp;
-
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 @Entity
@@ -45,7 +45,7 @@ public class SalesDeliveryNoteHdr {
     private String currencyCode;
 
     @Column(name = "CURRENCY_RATE")
-    private Long currencyRate;
+    private BigDecimal currencyRate;
 
     @Column(name = "DELIVERY_STATUS", length = 20)
     private String deliveryStatus;
